requirements.txt
django-embed-video==1.3.1
Django==2.2.6
Pillow==7.1.1
django-allauth==0.41.0
django-widget-tweaks==1.4.8
django-tinymce==3.0.1
whitenoise==5.0.1